% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olc_eval.R
\name{olc.eval}
\alias{olc.eval}
\title{Optimal Level Collapsing Evaluation}
\usage{
olc.eval(YY, Levels, model, k.max)
}
\arguments{
\item{YY}{Vector with response variable.}

\item{Levels}{Vector with the categorical variable to be collapsed.}

\item{model}{The model distribution to be used. Available families are: "binomial", "gaussian", "Gamma", "inverse.gaussian", "poisson", "quasi", "quasibinomial", "quasipoisson", "nb" (Negative Binomial), "poi_zero" (Zero Inflated Poisson), "neg_zero" (Zero Inflated Negative Binomial), "geo_zero" (Zero Inflated Geometric).}

\item{k.max}{The maximum number of groups to be evaluated.}
}
\value{
A list with the corresponded collapsed groups for different methods and number of collapsed groups.
}
\description{
Optimal Level Collapsing Evaluation
}
