% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olc.R
\name{olc}
\alias{olc}
\title{Optimal Level collapsing apply}
\usage{
olc(YY, Levels, model, k)
}
\arguments{
\item{YY}{Vector with response variable.}

\item{Levels}{Vector with the categorical variable to be collapsed.}

\item{model}{The model distribution to be used. See ?olc.eval for available model distributions.}

\item{k}{The number of groups to collapse the categorical variable.}
}
\value{
A list with the corresponded collapsed groups for different methods.
}
\description{
Optimal Level collapsing apply
}
