% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olc_AIC_plot.R
\name{olc_AIC_plot}
\alias{olc_AIC_plot}
\title{Optimal Level Collapsing AIC plot}
\usage{
olc_AIC_plot(object)
}
\arguments{
\item{object}{object created from olc.eval() function.}
}
\value{
A scatter plot with the AIC by number of collapsed groups from different methods.
}
\description{
Optimal Level Collapsing AIC plot
}
