% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olc_LR_plot.R
\name{olc_LR_plot}
\alias{olc_LR_plot}
\title{Optimal Level Collapsing LR-test p-value plot}
\usage{
olc_LR_plot(object)
}
\arguments{
\item{object}{object created from olc.eval() function.}
}
\value{
A scatter plot with the LR-test p-value by number of collapsed groups from different methods.
}
\description{
Optimal Level Collapsing LR-test p-value plot
}
